% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_importation.R
\name{importation}
\alias{importation}
\title{Import data associated with a list of sensors}
\usage{
importation(
  list_sensor_name,
  listeNom = c("Burel", "Leclerc", "ParisMarche", "rueVignes", "ParisArcEnCiel",
    "RteVitre", "RueGdDomaine", "StDidierNord", "rueVallee", "StDidierSud", "RuePrieure",
    "RueCottage", "RueVeronniere", "RueDesEcoles"),
  listeNombis = c("Burel-01", "Leclerc-02", "ParisMarche-03", "rueVignes-04",
    "ParisArcEnCiel-05", "RteVitre-06", "RueGdDomaine-07", "StDidierNord-08",
    "rueVallee-09", "StDidierSud-10", "RuePrieure-11", "RueCottage-12",
    "RueVeronniere-13", "RueDesEcoles-14")
)
}
\arguments{
\item{list_sensor_name}{A character vector specifying the names of sensors to import data for.}

\item{listeNom}{A character vector containing the name of each sensor that is displayed to the user}

\item{listeNombis}{A character vector containing the identifier name for each vector}
}
\value{
A data.frame containing the imported data.
}
\description{
This function imports data associated with a given list of sensor names.
}
