% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_write_update_data.R
\name{write_update_data}
\alias{write_update_data}
\title{Write or update the sensor data in the data folder}
\usage{
write_update_data(
  nom,
  date1,
  date2,
  listeNom = c("Burel", "Leclerc", "ParisMarche", "rueVignes", "ParisArcEnCiel",
    "RteVitre", "RueGdDomaine", "StDidierNord", "rueVallee", "StDidierSud", "RuePrieure",
    "RueCottage", "RueVeronniere", "RueDesEcoles"),
  listeNombis = c("Burel-01", "Leclerc-02", "ParisMarche-03", "rueVignes-04",
    "ParisArcEnCiel-05", "RteVitre-06", "RueGdDomaine-07", "StDidierNord-08",
    "rueVallee-09", "StDidierSud-10", "RuePrieure-11", "RueCottage-12",
    "RueVeronniere-13", "RueDesEcoles-14"),
  listeSegments = c(9000002156, 9000001906, 9000001618, 9000003090, 9000002453,
    9000001844, 9000001877, 9000002666, 9000002181, 9000002707, 9000003703, 9000003746,
    9000003775, 9000003736)
)
}
\arguments{
\item{nom}{Character. Name of the sensor.}

\item{date1}{Date. Start date in "yyyy-mm-dd" format.}

\item{date2}{Date. End date in "yyyy-mm-dd" format.}

\item{listeNom}{A character vector containing the name of each sensor that is displayed to the user}

\item{listeNombis}{A character vector containing the identifier name for each vector}

\item{listeSegments}{A character vector containing the ID of each vector}
}
\description{
This function writes or updates the sensor data in the data folder. It retrieves the data for the specified sensor between \code{date1} and \code{date2} (inclusive) using the \code{retrieve_sensor} function, and then converts certain columns to character strings before writing the data to a CSV file in the data folder.
}
